+++ src/main/kotlin/app/revanced/patches/youtube/layout/hidemixplaylists/patch/MixPlaylistsPatch.kt
@@ -5,13 +5,16 @@ import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.addInstruction
+import app.revanced.patcher.extensions.instruction
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import app.revanced.patcher.patch.PatchResult
 import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.patch.annotations.Patch
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patches.youtube.layout.hidemixplaylists.annotations.MixPlaylistsPatchCompatibility
-import app.revanced.patches.youtube.layout.hidemixplaylists.fingerprints.MixPlaylistsPatchFingerprint
+import app.revanced.patches.youtube.layout.hidemixplaylists.fingerprints.CreateMixPlaylistFingerprint
+import app.revanced.patches.youtube.layout.hidemixplaylists.fingerprints.SecondCreateMixPlaylistFingerprint
 import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
 import app.revanced.patches.youtube.misc.settings.bytecode.patch.SettingsPatch
 import app.revanced.patches.youtube.misc.settings.framework.components.impl.StringResource
@@ -21,12 +24,12 @@ import org.jf.dexlib2.iface.instruction.OneRegisterInstruction
 @Patch
 @DependsOn([IntegrationsPatch::class])
 @Name("hide-my-mix")
-@Description("Removes mix playlists from the feed.")
+@Description("Hides mix playlists.")
 @MixPlaylistsPatchCompatibility
 @Version("0.0.1")
 class MixPlaylistsPatch : BytecodePatch(
     listOf(
-        MixPlaylistsPatchFingerprint
+        CreateMixPlaylistFingerprint, SecondCreateMixPlaylistFingerprint
     )
 ) {
     override fun execute(context: BytecodeContext): PatchResult {
@@ -40,16 +43,22 @@ class MixPlaylistsPatch : BytecodePatch(
             )
         )
 
-        val result = MixPlaylistsPatchFingerprint.result!!
-        val method = result.mutableMethod
-        val index = result.scanResult.patternScanResult!!.endIndex - 6
-        val register = (method.implementation!!.instructions[index] as OneRegisterInstruction).registerA
-
-        method.addInstruction(
-            index + 2,
-            "invoke-static {v$register}, Lapp/revanced/integrations/patches/HideMixPlaylistsPatch;->hideMixPlaylists(Landroid/view/View;)V"
-        )
+        arrayOf(CreateMixPlaylistFingerprint, SecondCreateMixPlaylistFingerprint).forEach(::addHook)
 
         return PatchResultSuccess()
+    }
+
+    private fun addHook(fingerprint: MethodFingerprint) {
+        with (fingerprint.result!!) {
+            val insertIndex = scanResult.patternScanResult!!.endIndex - 3
+
+            val register = (mutableMethod.instruction(insertIndex - 2) as OneRegisterInstruction).registerA
+
+            mutableMethod.addInstruction(
+                insertIndex,
+                "invoke-static {v$register}, Lapp/revanced/integrations/patches/HideMixPlaylistsPatch;->hideMixPlaylists(Landroid/view/View;)V"
+            )
+        }
+
     }
 }