+++ src/main/kotlin/app/revanced/patches/youtube/layout/pivotbar/shortsbutton/patch/ShortsButtonRemoverPatch.kt
@@ -52,8 +52,8 @@ class ShortsButtonRemoverPatch : BytecodePatch(
             .onEach {
                 val resolutionSucceeded = it.resolve(
                     context,
-                    pivotBarResult.method,
-                    pivotBarResult.classDef
+                    pivotBarResult.mutableMethod,
+                    pivotBarResult.mutableClass
                 )
 
                 if (!resolutionSucceeded) return PatchResultError("${it.name} failed")
@@ -63,7 +63,7 @@ class ShortsButtonRemoverPatch : BytecodePatch(
         val enumScanResult = fingerprintResults[0]
         val buttonViewResult = fingerprintResults[1]
 
-        val enumHookInsertIndex = enumScanResult.startIndex
+        val enumHookInsertIndex = enumScanResult.startIndex + 2
         val buttonHookInsertIndex = buttonViewResult.endIndex
 
         /*